Implementation of file downloading in progress

+++ src/remotepatcher.h
@@ -13,8 +13,8 @@ public:
     virtual QString download(const QString& patcherSecret, const int& version) = 0;
 
 signals:
-    void bytesDownloadedChanged(const long& bytesDownloaded);
-    void totalBytesChanged(const long& totalBytes);
+    void bytesDownloadedChanged(const long long& bytesDownloaded);
+    void totalBytesChanged(const long long& totalBytes);
 };
 
 #endif // IREMOTEPATCHER_H
